Move EventForm component from features/event/components to features/event/components/EventForm
