+++ src/components/icnos.tsx
@@ -6,6 +6,7 @@ import {
   Copy,
   Check,
   CalendarCheck,
+  Loader2,
 } from 'lucide-react';
 
 export type Icon = LucideIcon;
@@ -17,4 +18,5 @@ export const Icons = {
   copy: Copy,
   check: Check,
   calendarCheck: CalendarCheck,
+  loaderC: Loader2,
 };
